automl/app/requirements_frozen.txt
pandas==2.2.2
ipython==8.25.0
Flask==3.0.2
google-cloud==0.34.0
google-cloud-storage==2.14.0
fsspec==2024.2.0
gcsfs==2024.2.0
google-cloud-pubsub==2.21.2
python-dotenv==1.0.1
Flask-MySQLdb==2.0.0
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.1
python-jose==3.3.0
Flask-JWT-Extended==4.6.0
## The following requirements were added by pip freeze:
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
Authlib==1.3.1
bcrypt==4.1.3
blinker==1.8.2
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==3.3.2
decorator==5.1.1
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.1.1
exceptiongroup==1.2.1
executing==2.0.1
Flask-Authlib==1.6.0
Flask-Bcrypt==1.0.1
Flask-JWT==0.3.2
Flask-Login==0.6.3
frozenlist==1.4.1
google-api-core==2.19.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-cloud-core==2.4.1
google-cloud-discoveryengine==0.11.10
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpc-google-iam-v1==0.13.0
grpcio==1.63.0
grpcio-status==1.62.2
idna==3.7
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jwt==1.3.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
multidict==6.0.5
mysqlclient==2.2.4
numpy==1.26.4
oauthlib==3.2.2
parso==0.8.4
pexpect==4.9.0
prompt_toolkit==3.0.46
proto-plus==1.23.0
protobuf==4.25.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.3
pydantic_core==2.18.4
Pygments==2.18.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.30
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.12.1
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
Werkzeug==3.0.3
yarl==1.9.4
